patch the inferno tsc plugin to emit named imports for the functions it needs rather than a full namespace import of Inferno
This still allows me to `import * as Inferno from 'inferno';` in my
modules when I need it. Turns out that the plugin in question was
emitting variable declarations like this immediately after the import:

    import * as Inferno from 'inferno';
    var createVNode = Inferno.createVNode;
    var createFragment = Inferno.createFragment;

And so on for every needed function, which kinda made the whole
namespace import thing redundant. What I did instead was to generate a
named import directly:

    import { createVNode, createFragment } from 'inferno';

It still makes those identifiers seem like global variables in my code
though, name conflicts are possible.