add a very basic key binding system, remove the window menu
Previously I used the OS'es built-in keyboard shortcut handling
mechanism by using the window menu, which worked well enough, but nw's
API for creating such menus is very limited, so it would be hard to
re-bind the keys after the menu is set up, for example. Not to mention
that on macOS a few of the "built-in" text editing shortcuts such as
Cmd+C and Cmd+V rely on the menu bar containing appropriate items, which
my application didn't implement. Rather than trying to use the limited
menu bar API, soon enough I'll move its items into the quick actions
menu, which is, of course, its primary purpose.